// src/enums/httpEnum.ts
/**
 * @description: Request result set
 */
export enum ResultEnum {
  SUCCESS = 1,
  ERROR = 0, // 请求异常
  TIMEOUT = 401,
  FAIL = -1, //请求失败
}

/**
 * @description: request method
 */
export enum RequestEnum {
  GET = 'GET',
  POST = 'POST',
  PUT = 'PUT',
  DELETE = 'DELETE',
}

/**
 * @description:  contentTyp
 */
export enum ContentTypeEnum {
  // json
  JSON = 'application/json;charset=UTF-8',
  // form-data qs
  FORM_URLENCODED = 'application/x-www-form-urlencoded;charset=UTF-8',
  // form-data  upload
  FORM_DATA = 'multipart/form-data;charset=UTF-8',
}